#22 [FEAT] oauth 관련 api 호출 방식 변경
