+++ api/oauth.ts
@@ -31,13 +31,26 @@ const kakaoOauth = {
   },
 };
 
-const googleOauth = {
-  getProfile: (accessToken: string) =>
-    oauthApi.get(`api/oauth2/google?token=${accessToken}`),
+const googleOauthAPI = {
+  getProfileNextAPI: (accessToken: string) =>
+    oauthApi.get(`api/auth/google?token=${accessToken}`),
+  getUserDataOrigin: (accessToken: string) =>
+    oauthApi.get(
+      `https://www.googleapis.com/oauth2/v3/userinfo?access_token=${accessToken}`
+    ),
 };
 
-// const naverOauth = {
-//   getProfile: () => {},
-// };
+const naverOauthAPI = {
+  getProfileNextAPI: (accessToken: string) => {
+    return oauthApi.get(`api/auth/naver?token=${accessToken}`);
+  },
+  getUserDataOrigin: (accessToken: string) => {
+    return oauthApi.get('https://openapi.naver.com/v1/nid/me', {
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+      },
+    });
+  },
+};
 
-export { kakaoOauth, googleOauth };
+export { kakaoOauth, googleOauthAPI, naverOauthAPI };
