+++ components/common/Oauth/GoogleOauth.tsx
@@ -1,10 +1,10 @@
+import Image from 'next/image';
 import styled from '@emotion/styled';
 import { GoogleOAuthProvider, useGoogleLogin } from '@react-oauth/google';
-import { googleOauth } from '@/api/oauth';
-import Image from 'next/image';
-import theme from '@/styles/theme';
 
+import theme from '@/styles/theme';
 import GoogleIconSVG from '@/public/google_icon.svg';
+import { googleOauthAPI } from '@/api/oauth';
 
 export const GoogleOauth = () => {
   return (
@@ -21,7 +21,7 @@ export const GoogleLoginButton = () => {
     onSuccess: async (tokenResponse) => {
       const accessToken = tokenResponse.access_token;
       if (accessToken) {
-        await googleOauth.getProfile(accessToken);
+        await googleOauthAPI.getProfileNextAPI(accessToken);
       }
     },
   });
